Validate Android build
Use stable Android SDK setup, fix gesture distance typing and accessibility settings intent, add command-model coverage, and retain diagnostic build logs.

+++ app/src/main/java/com/scrolldock/GestureFallback.kt
@@ -22,7 +22,7 @@ class GestureFallback(
         }
 
         val fraction = prefs.pagePercent / 100f
-        val distance = (usable * fraction).coerceAtMost(usable - service.dp(40))
+        val distance = (usable * fraction).coerceAtMost((usable - service.dp(40)).toFloat())
         var x = bounds.centerX().toFloat()
         if (overlay != null && overlay.contains(x.toInt(), bounds.centerY())) {
             x = bounds.left + bounds.width() * 0.35f
